+++ test_pipetools/test_utils.py
@@ -1,4 +1,4 @@
-from pipetools import X, sort_by, take_first, foreach, where, select_first
+from pipetools import X, sort_by, take_first, foreach, where, select_first, group_by
 from pipetools import unless, flatten, take_until, as_kwargs
 
 
@@ -20,6 +20,17 @@ class TestSortBy:
     def test_x(self):
 
         result = sort_by(-X[1])(zip('what', [1, 2, 3, 4]))
+
+        assert result == [
+            ('t', 4),
+            ('a', 3),
+            ('h', 2),
+            ('w', 1),
+        ]
+
+    def test_descending(self):
+
+        result = zip('what', [1, 2, 3, 4]) > sort_by(X[1]).descending
 
         assert result == [
             ('t', 4),
@@ -64,10 +75,10 @@ class TestSelectFirst:
 
     def test_select_first_none(self):
         result = select_first(X == 2)([0, 1, 0, 1])
-        assert result == None
+        assert result is None
 
     def test_select_first_empty(self):
-        assert select_first(X)([]) == None
+        assert select_first(X)([]) is None
 
 
 class TestAutoStringFormatter:
@@ -85,7 +96,7 @@ class TestUnless:
 
     def test_with_exception(self):
         f = unless(AttributeError, foreach(X.lower()) | list)
-        assert f(['A', 'B', 37]) == None
+        assert f(['A', 'B', 37]) is None
 
     def test_with_exception_in_foreach(self):
         f = foreach(unless(AttributeError, X.lower())) | list
@@ -97,7 +108,7 @@ class TestUnless:
 
     def test_partial_exc(self):
         f = unless(TypeError, enumerate, start=3)
-        assert f(42) == None
+        assert f(42) is None
 
     def test_X_ok(self):
         f = unless(TypeError, X * 'x')
@@ -105,7 +116,7 @@ class TestUnless:
 
     def test_X_exception(self):
         f = unless(TypeError, X * 'x')
-        assert f('x') == None
+        assert f('x') is None
 
 
 class TestFlatten:
@@ -153,3 +164,10 @@ class TestRegexCondidion:
             'foolproof',
         ]
         assert (data > select_first(r'^b.*r$')) == 'boo far'
+
+
+class TestGroupBy:
+
+    def test_basic(self):
+        src = [1, 2, 3, 4, 5, 6]
+        assert (src > group_by(X % 2) | dict) == {0: [2, 4, 6], 1: [1, 3, 5]}